fix(worker): Add delay to `tryToSetForeground` function
> The system will crash the app if the worker that calls setForeground() finished
> before the service runner be able to call Service.startForeground(). This edge
> case is not handled by WorkManager and there is no way to check if the required
> calls are done.

The delay duration is a magic number where we assume by then the
transition to foreground service is done.

Seems to be an issue introduced in Android 8, and Google refuses to fix
it, as usual.

REF: https://developer.android.com/about/versions/oreo/background.html#services (Last paragraph of the section)
REF: https://issuetracker.google.com/issues/76112072#comment36 (Status is set to Won't Fix)

Co-authored-by: Ivan Iskandar <12537387+ivaniskandar@users.noreply.github.com>

(cherry picked from commit 97bd9c4bc8ade8e3b795c5fbca0617ade347d03f)